bp_finder/requirements.txt
Django==1.10
djangorestframework==3.4.1
psycopg2==2.6.2